Colocate RolandIoSetup context and state
